+++ src/main/java/com/web01/animatch/servlet/RegistServlet.java
@@ -1,4 +1,4 @@
-package com.web01.animatch;
+package com.web01.animatch.servlet;
 
 import java.io.IOException;
 
@@ -6,17 +6,17 @@ import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpServletResponse;;
 
-public class LoginServlet extends HttpServlet {
+public class RegistServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
-    public LoginServlet() {
+    public RegistServlet() {
         super();
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String path = "/WEB-INF/jsp/login/login.jsp";
+		String path = "/WEB-INF/jsp/regist.jsp";
 		RequestDispatcher dispatcher = request.getRequestDispatcher(path);
 		dispatcher.forward(request, response);
 	}
@@ -24,5 +24,4 @@ public class LoginServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		doGet(request, response);
 	}
-
 }
