+++ src/main/java/com/web01/animatch/servlet/SearchTrimmerServlet.java
@@ -1,4 +1,4 @@
-package com.web01.animatch;
+package com.web01.animatch.servlet;
 
 import java.io.IOException;
 
@@ -6,17 +6,20 @@ import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;;
+import javax.servlet.http.HttpServletResponse;
 
-public class RegistServlet extends HttpServlet {
+public class SearchTrimmerServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
-    public RegistServlet() {
+    public SearchTrimmerServlet() {
         super();
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String path = "/WEB-INF/jsp/regist.jsp";
+		String path = "/WEB-INF/jsp/search/search_trimmer.jsp";
+		if(path.isEmpty()) {
+			return;
+		}
 		RequestDispatcher dispatcher = request.getRequestDispatcher(path);
 		dispatcher.forward(request, response);
 	}
