+++ src/main/webapp/WEB-INF/web.xml
@@ -7,67 +7,67 @@
     <servlet-name>IndexServlet</servlet-name>
     <display-name>IndexServlet</display-name>
     <description></description>
-  	<servlet-class>com.web01.animatch.IndexServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.IndexServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>LoginServlet</servlet-name>
   	<display-name>LoginServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.LoginServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.LoginServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>LogoutServlet</servlet-name>
   	<display-name>LogoutServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.LogoutServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.LogoutServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>RegistServlet</servlet-name>
   	<display-name>RegistServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.RegistServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.RegistServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>SearchOwnerServlet</servlet-name>
   	<display-name>SearchOwnerServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.SearchOwnerServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.SearchOwnerServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>SearchTrimmerServlet</servlet-name>
   	<display-name>SearchTrimmerServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.SearchTrimmerServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.SearchTrimmerServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>DetailOwnerServlet</servlet-name>
   	<display-name>DetailOwnerServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.DetailOwnerServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.DetailOwnerServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>DetailTrimmerServlet</servlet-name>
   	<display-name>DetailTrimmerServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.DetailTrimmerServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.DetailTrimmerServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>ContactInfServlet</servlet-name>
   	<display-name>ContactInfServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.ContactInfServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.ContactInfServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>MessageServlet</servlet-name>
   	<display-name>MessageServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.MessageServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.MessageServlet</servlet-class>
   </servlet>
   <servlet>
   	<servlet-name>ContactServlet</servlet-name>
   	<display-name>ContactServlet</display-name>
   	<description></description>
-  	<servlet-class>com.web01.animatch.ContactServlet</servlet-class>
+  	<servlet-class>com.web01.animatch.servlet.ContactServlet</servlet-class>
   </servlet>
   <servlet-mapping>
   	<servlet-name>IndexServlet</servlet-name>
